Fixed CRC compiler flag on int conversions

+++ Core/Src/fonts/glyph.c
@@ -16,6 +16,6 @@ void GetGlyphOutline(char glyph, PGlyphMetrics metric, PCBYTE *data) {
     // Let' s make sure our first glyph buffer is aligned at word boundary
     DebugAssert((((uintptr_t)s_glyphsData) & 0x3) == 0x0);
 
-	*metric = s_glyphs[glyph];
-	*data = s_glyphsData[glyph];
+	*metric = s_glyphs[(int)glyph];
+	*data = s_glyphsData[(int)glyph];
 }
